float64 -> float32

+++ panda_gym/pybullet.py
@@ -24,7 +24,7 @@ class PyBullet:
 
     def __init__(self, render: bool = False, n_substeps: int = 20, background_color: Optional[np.ndarray] = None) -> None:
         background_color = background_color if background_color is not None else np.array([223.0, 54.0, 45.0])
-        self.background_color = background_color.astype(np.float64) / 255
+        self.background_color = background_color.astype(np.float32) / 255
         options = "--background_color_red={} \
                     --background_color_green={} \
                     --background_color_blue={}".format(
